refactor(settings): extract AboutSection (and SoftwareUpdates) to sections/
Moves the About tab content — app version, software update check/install,
and the keyboard-shortcut reference — into its own file alongside the
other section components under src/components/settings/sections/.

No behavior changes.
